+++ Common_glTF_Exporter/Utils/GLTFBinaryDataUtils.cs
@@ -112,7 +112,7 @@
             return byteOffset + vec3ViewNormals.byteLength;
         }
 
-        public static int ExportBatchId(int bufferIdx, int byteOffset, int sizeOfVec3View, int elementsPerVertex, int elementId, GeometryDataObject geomData, GLTFBinaryData bufferData, List<GLTFBufferView> bufferViews, List<GLTFAccessor> accessors)
+        public static int ExportBatchId(int bufferIdx, int byteOffset, int sizeOfVec3View, int elementsPerVertex, long elementId, GeometryDataObject geomData, GLTFBinaryData bufferData, List<GLTFBufferView> bufferViews, List<GLTFAccessor> accessors)
         {
             for (int i = 0; i < geomData.Vertices.Count; i++)
             {
